requirements.txt
tensorflow>=2.8.0
opencv-python>=4.5.5
numpy>=1.21.0
matplotlib>=3.5.0
Pillow>=9.0.0